requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.0
google-generativeai==0.8.0
python-dotenv==1.0.1
jinja2==3.1.4
